requirements.txt
asgiref==3.6.0
boto3==1.26.133
botocore==1.29.133
dj-database-url==2.0.0
Django==4.2.1
django-bootstrap-form==3.4
django-environ==0.10.0
django-filter==23.2
django-heroku==0.3.1
gunicorn==20.1.0
jmespath==1.0.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
